+++ benches/simd.rs
@@ -12,7 +12,7 @@ pub fn benchmark_simd_integrate(c: &mut Criterion) {
         max_z: 30.,
         sphere_radius: 1.,
     };
-    let max_velocity: f64 = 20.;
+    let max_velocity: f32 = 20.;
     let num_spheres = 350000;
     let physics_spheres = physics::get_random_physics_data(
         &mut generator,
